requirements.txt
llama2
Pillow==10.2.0
streamlit==1.32.2
torch==2.2.1
transformers==4.40.0.dev0